separate building from testing
