+++ package.json
@@ -6,8 +6,8 @@
   "license": "MIT",
   "repository": "https://github.com/gobblejs/gobble",
   "dependencies": {
-    "buffer-crc32": "~0.2.3",
     "chalk": "^1.0.0",
+    "crc": "^3.2.1",
     "debounce": "^1.0.0",
     "eventemitter2": "^0.4.14",
     "findup-sync": "~0.1.3",
@@ -22,7 +22,7 @@
     "resolve": "^1.0.0",
     "rimraf": "~2.2.8",
     "sander": "^0.3.2",
-    "sorcery": "^0.5.4",
+    "sorcery": "^0.5.5",
     "source-map-support": "^0.2.10",
     "tiny-lr": "~0.1.0"
   },
@@ -43,8 +43,9 @@
     "source-map": "^0.4.2"
   },
   "scripts": {
-    "prepublish": "set -e; npm run build",
-    "test": "sh ./scripts/test.sh",
-    "build": "set -e; npm test; rm -rf lib; cp -r tmp lib"
+    "prepublish": "npm test",
+    "test": "mocha test/test.js",
+    "pretest": "npm run build",
+    "build": "sh ./scripts/build.sh"
   }
 }
