+++ scripts/build.sh
@@ -2,6 +2,8 @@
 
 set -e
 
+rm -rf lib
+
 echo "babel..."
 babel src --out-dir .babel -b es6.modules,useStrict --source-maps-inline --loose es6.classes > /dev/null
 
@@ -9,10 +11,7 @@ echo "esperanto..."
 esperanto -i .babel -o .esperanto -m inline -t cjs -s
 
 echo "sorcery"
-sorcery -i .esperanto -o tmp
-
-echo "mocha..."
-mocha test/test.js
+sorcery -i .esperanto -o lib
 
 rm -rf .babel
 rm -rf .esperanto
