+++ test/sample/foo.js
@@ -1,4 +1,4 @@
-var gobble = require( '../../tmp' ).default,
+var gobble = require( '../../lib' ).default,
 	path = require( 'path' );
 
 module.exports = gobble( path.join( __dirname, 'foo' ) );
